Split app pane and MCP runtime boundaries
