+++ Sources/QuillCodeApp/QuillCodeBrowserPaneView.swift
@@ -25,234 +25,6 @@ struct QuillCodeBrowserPaneView: View {
         .background(QuillCodePalette.panel)
     }
 
-    private var header: some View {
-        HStack(spacing: QuillCodeMetrics.controlClusterSpacing) {
-            Image(systemName: "globe")
-                .foregroundStyle(QuillCodePalette.blue)
-            Text("Browser")
-                .font(.headline)
-            Text(browser.statusLabel)
-                .font(.caption)
-                .foregroundStyle(QuillCodePalette.muted)
-            Spacer()
-        }
-    }
-
-    private var tabStrip: some View {
-        ScrollView(.horizontal, showsIndicators: false) {
-            HStack(spacing: QuillCodeMetrics.controlClusterSpacing) {
-                ForEach(browser.tabs) { tab in
-                    Button {
-                        onCommand(tab.selectCommandID)
-                    } label: {
-                        HStack(spacing: QuillCodeMetrics.denseControlClusterSpacing) {
-                            Text(tab.title)
-                                .font(.caption.weight(tab.isActive ? .semibold : .regular))
-                                .lineLimit(1)
-                            if let urlLabel = tab.urlLabel, tab.isActive {
-                                Text(urlLabel)
-                                    .font(.caption2)
-                                    .foregroundStyle(QuillCodePalette.muted)
-                                    .lineLimit(1)
-                            }
-                        }
-                        .quillCodeCapsuleButtonTarget(minWidth: 112)
-                    }
-                    .buttonStyle(QuillCodePressableButtonStyle())
-                    .foregroundStyle(tab.isActive ? QuillCodePalette.text : QuillCodePalette.muted)
-                    .background(tab.isActive ? QuillCodePalette.selection.opacity(0.42) : QuillCodePalette.panel.opacity(0.8))
-                    .clipShape(Capsule())
-                    .accessibilityLabel(tab.isActive ? "Current browser tab \(tab.title)" : "Switch to browser tab \(tab.title)")
-                }
-
-                browserNavigationButton(
-                    systemName: "plus",
-                    label: "New browser tab",
-                    isEnabled: true
-                ) {
-                    onCommand("browser-tab-new")
-                }
-
-                browserNavigationButton(
-                    systemName: "xmark",
-                    label: "Close browser tab",
-                    isEnabled: browser.canCloseActiveTab
-                ) {
-                    onCommand("browser-tab-close:\(browser.activeTabID.uuidString)")
-                }
-            }
-        }
-    }
-
-    private var navigationBar: some View {
-        ViewThatFits(in: .horizontal) {
-            horizontalNavigationBar
-            compactNavigationBar
-        }
-    }
-
-    private var horizontalNavigationBar: some View {
-        HStack(spacing: QuillCodeMetrics.controlClusterSpacing) {
-            browserNavigationControls
-            browserAddressControls
-        }
-    }
-
-    private var compactNavigationBar: some View {
-        VStack(alignment: .leading, spacing: 8) {
-            browserNavigationControls
-            browserAddressControls
-        }
-    }
-
-    private var browserNavigationControls: some View {
-        HStack(spacing: QuillCodeMetrics.controlClusterSpacing) {
-            browserNavigationButton(
-                systemName: "chevron.left",
-                label: "Back",
-                isEnabled: browser.canGoBack
-            ) {
-                onCommand("browser-back")
-            }
-            browserNavigationButton(
-                systemName: "chevron.right",
-                label: "Forward",
-                isEnabled: browser.canGoForward
-            ) {
-                onCommand("browser-forward")
-            }
-            browserNavigationButton(
-                systemName: "arrow.clockwise",
-                label: "Reload",
-                isEnabled: browser.canReload
-            ) {
-                onCommand("browser-reload")
-            }
-            if let onOpenSession {
-                Button("Session", action: onOpenSession)
-                    .buttonStyle(QuillCodePressableButtonStyle())
-                    .quillCodeTextButtonTarget(minWidth: 84)
-                    .disabled(!browser.canOpen && browser.currentURL == nil)
-                    .help("Open a visible browser session using QuillCode's persistent browser profile.")
-                    .accessibilityIdentifier("quillcode-browser-session")
-            }
-        }
-    }
-
-    private var browserAddressControls: some View {
-        HStack(spacing: QuillCodeMetrics.controlClusterSpacing) {
-            TextField("localhost:3000, docs/page.html, or https://example.com", text: $addressDraft)
-                .textFieldStyle(.roundedBorder)
-                .onSubmit(onOpen)
-                .quillCodeTextEntryTarget()
-                .frame(maxWidth: .infinity)
-                .accessibilityIdentifier("quillcode-browser-address")
-            Button("Open", action: onOpen)
-                .buttonStyle(QuillCodePressableButtonStyle())
-                .quillCodeTextButtonTarget()
-                .disabled(addressDraft.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty)
-                .accessibilityIdentifier("quillcode-browser-action")
-        }
-    }
-
-    @ViewBuilder
-    private var pageSummary: some View {
-        if let currentURL = browser.currentURL {
-            currentPageSummary(currentURL: currentURL)
-        } else {
-            emptyPageSummary
-        }
-    }
-
-    private func currentPageSummary(currentURL: String) -> some View {
-        VStack(alignment: .leading, spacing: 6) {
-            HStack(spacing: QuillCodeMetrics.controlClusterSpacing) {
-                Text(browser.title)
-                    .font(.callout.weight(.semibold))
-                    .lineLimit(1)
-                if let snapshot = browser.snapshot {
-                    browserBadge(snapshot.sourceLabel, tint: QuillCodePalette.blue)
-                    browserBadge(
-                        snapshot.inspectionDepthLabel,
-                        tint: browserInspectionTint(snapshot.inspectionDepth)
-                    )
-                }
-            }
-            Text(currentURL)
-                .font(.caption.monospaced())
-                .foregroundStyle(QuillCodePalette.muted)
-                .lineLimit(1)
-            snapshotSummary
-        }
-        .padding(10)
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .quillCodeSurface(
-            fill: QuillCodePalette.background.opacity(0.7),
-            radius: 20,
-            stroke: Color.white.opacity(0.08),
-            shadow: false
-        )
-    }
-
-    @ViewBuilder
-    private var snapshotSummary: some View {
-        if let snapshot = browser.snapshot {
-            Text(snapshot.summary)
-                .font(.caption)
-                .foregroundStyle(QuillCodePalette.text)
-            LazyVGrid(columns: [GridItem(.adaptive(minimum: 150), spacing: 6)], alignment: .leading, spacing: 6) {
-                ForEach(snapshot.details, id: \.self) { detail in
-                    Text(detail)
-                        .font(.caption2.monospaced())
-                        .foregroundStyle(QuillCodePalette.muted)
-                        .lineLimit(1)
-                        .padding(.horizontal, 7)
-                        .padding(.vertical, 4)
-                        .background(QuillCodePalette.panel.opacity(0.9))
-                        .clipShape(Capsule())
-                }
-            }
-            if !snapshot.outline.isEmpty {
-                pageOutline(snapshot.outline)
-            }
-            if let textSnippet = snapshot.textSnippet {
-                Text(textSnippet)
-                    .font(.caption2)
-                    .foregroundStyle(QuillCodePalette.muted)
-                    .lineLimit(4)
-            }
-        } else {
-            Text("Ready for page inspection.")
-                .font(.caption)
-                .foregroundStyle(QuillCodePalette.muted)
-        }
-    }
-
-    private func pageOutline(_ outline: [String]) -> some View {
-        VStack(alignment: .leading, spacing: 3) {
-            Text("Page outline")
-                .font(.caption2.weight(.semibold))
-                .foregroundStyle(QuillCodePalette.muted)
-            ForEach(outline.prefix(8), id: \.self) { item in
-                Text(item)
-                    .font(.caption2)
-                    .foregroundStyle(QuillCodePalette.text)
-                    .lineLimit(1)
-            }
-        }
-    }
-
-    private var emptyPageSummary: some View {
-        VStack(alignment: .leading, spacing: 4) {
-            Text(browser.emptyTitle)
-                .font(.callout.weight(.semibold))
-            Text(browser.emptySubtitle)
-                .font(.caption)
-                .foregroundStyle(QuillCodePalette.muted)
-        }
-        .frame(maxWidth: .infinity, alignment: .leading)
-    }
-
     private var commentInput: some View {
         HStack(spacing: QuillCodeMetrics.controlClusterSpacing) {
             TextField("Add browser comment", text: $commentDraft)
@@ -298,57 +70,10 @@ struct QuillCodeBrowserPaneView: View {
         }
     }
 
-    private func browserNavigationButton(
-        systemName: String,
-        label: String,
-        isEnabled: Bool,
-        action: @escaping () -> Void
-    ) -> some View {
-        Button(action: action) {
-            Image(systemName: systemName)
-                .quillCodeIconButtonTarget()
-                .foregroundStyle(isEnabled ? QuillCodePalette.blue : QuillCodePalette.muted.opacity(0.55))
-                .background(QuillCodePalette.selection.opacity(isEnabled ? 0.34 : 0.16))
-                .clipShape(RoundedRectangle(cornerRadius: QuillCodeMetrics.iconControlRadius, style: .continuous))
-        }
-        .buttonStyle(QuillCodePressableButtonStyle())
-        .disabled(!isEnabled)
-        .accessibilityLabel(label)
-    }
-
     private func addComment() {
         let comment = commentDraft.trimmingCharacters(in: .whitespacesAndNewlines)
         guard !comment.isEmpty else { return }
         onAddComment(comment)
         commentDraft = ""
-    }
-
-    private func browserBadge(_ label: String, tint: Color) -> some View {
-        Text(label)
-            .font(.caption2.monospacedDigit().weight(.bold))
-            .foregroundStyle(tint)
-            .padding(.horizontal, 7)
-            .padding(.vertical, 3)
-            .background(tint.opacity(0.14))
-            .overlay(
-                Capsule()
-                    .stroke(tint.opacity(0.22), lineWidth: 1)
-            )
-            .clipShape(Capsule())
-    }
-
-    private func browserInspectionTint(_ depth: BrowserInspectionDepth) -> Color {
-        switch depth {
-        case .metadataOnly:
-            return QuillCodePalette.yellow
-        case .fileMetadata:
-            return QuillCodePalette.blue
-        case .staticHTMLSnapshot:
-            return QuillCodePalette.green
-        case .networkHTMLSnapshot:
-            return QuillCodePalette.green
-        case .liveDOMSnapshot:
-            return QuillCodePalette.purple
-        }
     }
 }
