+++ Sources/QuillCodeApp/WorkspaceMCPRuntime.swift
@@ -224,130 +224,11 @@ final class WorkspaceMCPRuntime: @unchecked Sendable {
         )
     }
 
-    static func executionOverride(
-        sessions: [String: any WorkspaceMCPSession],
-        summaries: [String: MCPServerProbeSummary]
-    ) -> AgentToolExecutionOverride? {
-        guard !sessions.isEmpty else { return nil }
-
-        return { call, _ in
-            Self.execute(call: call, sessions: sessions, summaries: summaries)
-        }
-    }
-
-    private static func execute(
-        call: ToolCall,
-        sessions: [String: any WorkspaceMCPSession],
-        summaries: [String: MCPServerProbeSummary]
-    ) -> ToolResult? {
-        let allowedTools = summaries.mapValues { Set($0.toolNames) }
-        let allowedPrompts = summaries.mapValues { Set($0.promptNames) }
-
-        do {
-            switch call.name {
-            case ToolDefinition.mcpCall.name:
-                let request = try MCPToolCallRequest(argumentsJSON: call.argumentsJSON)
-                guard let session = sessions[request.serverID] else {
-                    return ToolResult(ok: false, error: "MCP server is not running or is not Ready: \(request.serverID)")
-                }
-                guard allowedTools[request.serverID]?.contains(request.toolName) == true else {
-                    return ToolResult(
-                        ok: false,
-                        error: "MCP tool \(request.toolName) was not advertised by \(request.serverID)."
-                    )
-                }
-                return try session.callTool(
-                    toolName: request.toolName,
-                    argumentsJSON: request.toolArgumentsJSON,
-                    timeout: 10.0
-                )
-
-            case ToolDefinition.mcpReadResource.name:
-                let request = try MCPResourceReadRequest(argumentsJSON: call.argumentsJSON)
-                guard let session = sessions[request.serverID] else {
-                    return ToolResult(ok: false, error: "MCP server is not running or is not Ready: \(request.serverID)")
-                }
-                guard let uri = request.resourceURI(in: summaries[request.serverID]) else {
-                    return ToolResult(
-                        ok: false,
-                        error: "MCP resource \(request.resourceIdentifier) was not advertised by \(request.serverID)."
-                    )
-                }
-                return try session.readResource(uri: uri, timeout: 10.0)
-
-            case ToolDefinition.mcpGetPrompt.name:
-                let request = try MCPPromptGetRequest(argumentsJSON: call.argumentsJSON)
-                guard let session = sessions[request.serverID] else {
-                    return ToolResult(ok: false, error: "MCP server is not running or is not Ready: \(request.serverID)")
-                }
-                guard allowedPrompts[request.serverID]?.contains(request.promptName) == true else {
-                    return ToolResult(
-                        ok: false,
-                        error: "MCP prompt \(request.promptName) was not advertised by \(request.serverID)."
-                    )
-                }
-                return try session.getPrompt(
-                    name: request.promptName,
-                    argumentsJSON: request.promptArgumentsJSON,
-                    timeout: 10.0
-                )
-
-            default:
-                return nil
-            }
-        } catch {
-            return ToolResult(ok: false, error: Self.userFacingError(error))
-        }
-    }
-
-    static func probeNoticeSuffix(for result: MCPServerProbeResult) -> String {
-        let toolPreview = result.toolNames.prefix(3).joined(separator: ", ")
-        let toolLabel: String
-        if result.toolNames.isEmpty {
-            toolLabel = "0 tools"
-        } else {
-            let remaining = result.toolNames.count - min(result.toolNames.count, 3)
-            let noun = result.toolNames.count == 1 ? "tool" : "tools"
-            toolLabel = remaining > 0
-                ? "\(result.toolNames.count) \(noun): \(toolPreview), +\(remaining) more"
-                : "\(result.toolNames.count) \(noun): \(toolPreview)"
-        }
-        let resourceLabel = result.resourceNames.isEmpty
-            ? nil
-            : "\(result.resourceNames.count) resource\(result.resourceNames.count == 1 ? "" : "s")"
-        let promptLabel = result.promptNames.isEmpty
-            ? nil
-            : "\(result.promptNames.count) prompt\(result.promptNames.count == 1 ? "" : "s")"
-        let parts = [toolLabel, resourceLabel, promptLabel].compactMap { $0 }
-        return " (\(parts.joined(separator: "; ")))"
-    }
-
     private func stopProcess(id: String) {
         if let handle = processes[id], handle.process.isRunning {
             handle.process.clearReadabilityHandlers()
             handle.process.terminate()
         }
         processes[id] = nil
-    }
-
-    private static func userFacingError(_ error: Error) -> String {
-        if let localized = (error as? LocalizedError)?.errorDescription,
-           !localized.isEmpty {
-            return localized
-        }
-        return String(describing: error)
-    }
-}
-
-private final class WorkspaceMCPProcessHandle: @unchecked Sendable {
-    let process: any WorkspaceMCPProcessControlling
-    let session: any WorkspaceMCPSession
-
-    init(
-        process: any WorkspaceMCPProcessControlling,
-        session: any WorkspaceMCPSession
-    ) {
-        self.process = process
-        self.session = session
     }
 }
