+++ Tests/QuillCodeParityTests/ParityMCPGateTests.swift
@@ -7,6 +7,9 @@ final class ParityMCPGateTests: QuillCodeParityTestCase {
         let mcpModelText = try Self.appSourceText(named: "WorkspaceModelMCP.swift")
         let mcpRequestText = try Self.appSourceText(named: "WorkspaceMCPRequests.swift")
         let mcpRuntimeText = try Self.appSourceText(named: "WorkspaceMCPRuntime.swift")
+        let mcpRuntimeExecutionText = try Self.appSourceText(named: "WorkspaceMCPRuntimeExecution.swift")
+        let mcpProcessHandleText = try Self.appSourceText(named: "WorkspaceMCPProcessHandle.swift")
+        let mcpProbeNoticeText = try Self.appSourceText(named: "WorkspaceMCPRuntimeProbeNotice.swift")
         let mcpLauncherText = try Self.appSourceText(named: "WorkspaceMCPServerLauncher.swift")
         let mcpCatalogText = try Self.appSourceText(named: "WorkspaceMCPToolCatalog.swift")
 
@@ -17,7 +20,9 @@ final class ParityMCPGateTests: QuillCodeParityTestCase {
         XCTAssertTrue(mcpRequestText.contains("struct MCPResourceReadRequest"), "MCP resource parsing should live in a focused request parser file.")
         XCTAssertTrue(mcpRequestText.contains("struct MCPPromptGetRequest"), "MCP prompt parsing should live in a focused request parser file.")
         XCTAssertTrue(mcpRuntimeText.contains("final class WorkspaceMCPRuntime"), "MCP process lifecycle should live in a focused runtime file.")
-        XCTAssertTrue(mcpRuntimeText.contains("private final class WorkspaceMCPProcessHandle"), "MCP process handles should be private to the runtime.")
+        XCTAssertTrue(mcpProcessHandleText.contains("final class WorkspaceMCPProcessHandle"), "MCP process handles should live in a focused runtime support file.")
+        XCTAssertTrue(mcpProcessHandleText.contains("WorkspaceMCPProcessControlling"), "MCP process handles should keep process abstraction typed.")
+        XCTAssertTrue(mcpProcessHandleText.contains("WorkspaceMCPSession"), "MCP process handles should keep session abstraction typed.")
         XCTAssertTrue(mcpLauncherText.contains("protocol WorkspaceMCPServerLaunching"), "MCP process launch should have an injectable launcher protocol.")
         XCTAssertTrue(mcpLauncherText.contains("struct WorkspaceMCPLaunchRequest"), "MCP launch request validation should live beside the launcher.")
         XCTAssertTrue(mcpLauncherText.contains("struct WorkspaceMCPProcessLaunchConfiguration"), "MCP command resolution should live beside the launcher.")
@@ -27,7 +32,10 @@ final class ParityMCPGateTests: QuillCodeParityTestCase {
         XCTAssertTrue(mcpRuntimeText.contains("launcher.launch("), "MCP runtime should delegate process creation to the launcher.")
         XCTAssertTrue(mcpCatalogText.contains("struct WorkspaceMCPToolCatalog"), "MCP dynamic tool descriptions should live in a focused catalog file.")
         XCTAssertTrue(mcpRuntimeText.contains("WorkspaceMCPToolCatalog("), "MCP runtime should delegate dynamic tool definitions to the catalog.")
-        XCTAssertTrue(mcpRuntimeText.contains("static func executionOverride"), "MCP dynamic tool routing should live in the runtime.")
+        XCTAssertTrue(mcpRuntimeExecutionText.contains("static func executionOverride"), "MCP dynamic tool routing should live in focused runtime execution support.")
+        XCTAssertTrue(mcpRuntimeExecutionText.contains("executeToolCall"), "MCP tool execution should be named separately from prompt and resource execution.")
+        XCTAssertTrue(mcpRuntimeExecutionText.contains("WorkspaceMCPAdvertisedCapabilities"), "MCP execution should centralize advertised-capability checks.")
+        XCTAssertTrue(mcpProbeNoticeText.contains("static func probeNoticeSuffix"), "MCP probe summary copy should live outside process lifecycle.")
         XCTAssertTrue(mcpModelText.contains("extension QuillCodeWorkspaceModel"), "MCP workspace APIs should live in a focused model extension.")
         XCTAssertTrue(mcpModelText.contains("func startMCPServer"), "MCP start orchestration should live in the focused model extension.")
         XCTAssertTrue(mcpModelText.contains("func stopMCPServer"), "MCP stop orchestration should live in the focused model extension.")
@@ -46,6 +54,8 @@ final class ParityMCPGateTests: QuillCodeParityTestCase {
         XCTAssertFalse(mcpRuntimeText.contains("Process()"), "WorkspaceMCPRuntime should not construct concrete processes directly.")
         XCTAssertFalse(mcpRuntimeText.contains("MCPStdioProber("), "WorkspaceMCPRuntime should not construct stdio sessions directly.")
         XCTAssertFalse(mcpRuntimeText.contains("URL(fileURLWithPath: \"/usr/bin/env\")"), "WorkspaceMCPRuntime should not resolve launch commands directly.")
+        XCTAssertFalse(mcpRuntimeText.contains("private static func execute"), "WorkspaceMCPRuntime should not own dynamic tool routing details.")
+        XCTAssertFalse(mcpRuntimeText.contains("private static func userFacingError"), "WorkspaceMCPRuntime should not own dynamic tool error formatting.")
         XCTAssertTrue(mcpLauncherText.contains("Process()"), "Concrete process construction should be isolated to the MCP launcher.")
         XCTAssertFalse(modelText.contains("readyMCPToolDescriptions"), "WorkspaceModel should not format MCP tool descriptions directly.")
         XCTAssertFalse(mcpRuntimeText.contains("func readyToolDescriptions"), "MCP runtime should not format MCP tool descriptions directly.")
