+++ Tests/QuillCodeParityTests/ParityNativeInteractionContractGateTests.swift
@@ -420,6 +420,7 @@ final class ParityNativeInteractionContractGateTests: QuillCodeParityTestCase {
         let sidebarText = try Self.appSourceText(named: "QuillCodeSidebarView.swift")
         let terminalText = try Self.appSourceText(named: "QuillCodeTerminalPaneView.swift")
         let browserText = try Self.appSourceText(named: "QuillCodeBrowserPaneView.swift")
+        let browserControlsText = try Self.appSourceText(named: "QuillCodeBrowserPaneControls.swift")
         let automationCreateMenuText = try Self.appSourceText(named: "QuillCodeAutomationCreateMenu.swift")
 
         XCTAssertTrue(
@@ -443,8 +444,8 @@ final class ParityNativeInteractionContractGateTests: QuillCodeParityTestCase {
         XCTAssertTrue(
             terminalText.contains(#".accessibilityIdentifier("quillcode-terminal-command")"#)
                 && terminalText.contains(#".accessibilityIdentifier("quillcode-terminal-action")"#)
-                && browserText.contains(#".accessibilityIdentifier("quillcode-browser-address")"#)
-                && browserText.contains(#".accessibilityIdentifier("quillcode-browser-action")"#)
+                && browserControlsText.contains(#".accessibilityIdentifier("quillcode-browser-address")"#)
+                && browserControlsText.contains(#".accessibilityIdentifier("quillcode-browser-action")"#)
                 && browserText.contains(#".accessibilityIdentifier("quillcode-browser-add-comment")"#),
             "Terminal and browser input/action controls should expose stable native accessibility IDs."
         )
