+++ Tests/QuillCodeParityTests/ParityWorkspaceSurfaceGateTests.swift
@@ -38,6 +38,8 @@ final class ParityWorkspaceSurfaceGateTests: QuillCodeParityTestCase {
         let mainPaneText = try Self.appSourceText(named: "QuillCodeWorkspaceMainPaneView.swift")
         let chromeText = try Self.appSourceText(named: "QuillCodeSecondaryPanesView.swift")
         let extensionsText = try Self.appSourceText(named: "QuillCodeExtensionsPaneView.swift")
+        let extensionCardsText = try Self.appSourceText(named: "QuillCodeExtensionsPaneCards.swift")
+        let extensionMetadataText = try Self.appSourceText(named: "QuillCodeExtensionsPaneProbeMetadata.swift")
         let memoriesText = try Self.appSourceText(named: "QuillCodeMemoriesPaneView.swift")
         let automationsText = try Self.appSourceText(named: "QuillCodeAutomationsPaneView.swift")
         let automationMenuText = try Self.appSourceText(named: "QuillCodeAutomationCreateMenu.swift")
@@ -47,7 +49,10 @@ final class ParityWorkspaceSurfaceGateTests: QuillCodeParityTestCase {
         XCTAssertTrue(chromeText.contains("struct QuillCodePaneCountPill"), "Secondary pane count pills should remain shared native chrome.")
         XCTAssertTrue(chromeText.contains("struct QuillCodePaneEmptyStateView"), "Secondary pane empty states should remain shared native chrome.")
         XCTAssertTrue(extensionsText.contains("struct QuillCodeExtensionsPaneView"), "Extensions native UI should live in its own focused file.")
-        XCTAssertTrue(extensionsText.contains("ProjectExtensionManifestSurface"), "MCP extension metadata display should stay with the Extensions native pane.")
+        XCTAssertTrue(extensionCardsText.contains("ProjectExtensionManifestSurface"), "Extension cards should own manifest row display.")
+        XCTAssertTrue(extensionCardsText.contains("extensionCommand(id:"), "Extension cards should own row action command construction.")
+        XCTAssertTrue(extensionMetadataText.contains("MCPToolDescriptor"), "MCP extension metadata display should stay in focused Extensions metadata UI.")
+        XCTAssertTrue(extensionMetadataText.contains("probeReferenceActionGroup"), "MCP resource and prompt actions should be grouped in focused metadata UI.")
         XCTAssertTrue(memoriesText.contains("struct QuillCodeMemoriesPaneView"), "Memories native UI should live in its own focused file.")
         XCTAssertTrue(automationsText.contains("struct QuillCodeAutomationsPaneView"), "Automations native UI should live in its own focused file.")
         XCTAssertTrue(automationsText.contains("QuillCodeAutomationCreateMenu"), "Automations pane should delegate create-menu routing.")
@@ -65,6 +70,8 @@ final class ParityWorkspaceSurfaceGateTests: QuillCodeParityTestCase {
         XCTAssertFalse(chromeText.contains("struct QuillCodeExtensionsPaneView"), "Shared secondary chrome should not own Extensions pane content.")
         XCTAssertFalse(chromeText.contains("struct QuillCodeMemoriesPaneView"), "Shared secondary chrome should not own Memories pane content.")
         XCTAssertFalse(chromeText.contains("struct QuillCodeAutomationsPaneView"), "Shared secondary chrome should not own Automations pane content.")
+        XCTAssertFalse(extensionsText.contains("private func extensionCard"), "Extensions pane shell should not own extension-card internals.")
+        XCTAssertFalse(extensionsText.contains("probeMetadataChips"), "Extensions pane shell should not own MCP metadata internals.")
         XCTAssertFalse(automationsText.contains("private var createMenu"), "Automations pane shell should not own create-menu internals.")
         XCTAssertFalse(automationsText.contains("private func automationCard"), "Automations pane shell should not own workflow card internals.")
         XCTAssertFalse(automationsText.contains("automationCommand(id:"), "Automations pane shell should not own row action command construction.")
